+++ phpunit/tests/test-post-index-helpers.php
@@ -319,4 +319,21 @@ class Post_Index_Helpers_Test extends WP_UnitTestCase {
 		$this->expectOutputRegex( '/^$/', c2c_the_index() );
 	}
 
+	/* action: c2c_the_index() */
+
+	public function test_action_c2c_the_index_for_valid_index() {
+		$this->loop( 3 );
+
+		$this->expectOutputRegex( '/^2$/', do_action( 'c2c_the_index' ) );
+	}
+
+	public function test_action_c2c_the_index_for_valid_index_of_0() {
+		$this->loop();
+
+		$this->expectOutputRegex( '/^0$/', do_action( 'c2c_the_index' ) );
+	}
+
+	public function test_action_c2c_the_index_for_invalid_index() {
+		$this->expectOutputRegex( '/^$/', do_action( 'c2c_the_index' ) );
+	}
 }